# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for managing recipes.",
        "title": "Recipes API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/recipes": {
            "get": {
                "description": "Returns list of recipes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "List Recipes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListRecipes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Create a new recipe",
                "parameters": [
                    {
                        "description": "Add Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddUpdateRecipe"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/recipes/search": {
            "get": {
                "description": "Search recipes by tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Search recipes by tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag to search for",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListRecipes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/recipes/{id}": {
            "get": {
                "description": "Returns a single recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Get a recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ViewRecipe"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Update a recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Recipe",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddUpdateRecipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Delete a recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddUpdateRecipe": {
            "type": "object",
            "required": [
                "ingredients",
                "instructions",
                "name",
                "tags"
            ],
            "properties": {
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "2 1/4 cups all-purpose flour",
                        "1 tsp baking soda",
                        "1 cup butter",
                        "3/4 cup granulated sugar",
                        "3/4 cup brown sugar",
                        "2 large eggs",
                        "2 cups semi-sweet chocolate chips"
                    ]
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Preheat oven to 375°F (190°C)",
                        "Mix dry ingredients",
                        "Cream butter and sugars",
                        "Beat in eggs",
                        "Stir in chocolate chips",
                        "Drop spoonfuls onto baking sheets",
                        "Bake for 9 to 11 minutes"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "Chocolate Chip Cookies"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "dessert",
                        "snack"
                    ]
                }
            }
        },
        "models.ListRecipes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ViewRecipe"
                    }
                }
            }
        },
        "models.ViewRecipe": {
            "type": "object",
            "required": [
                "ingredients",
                "instructions",
                "name",
                "tags"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "c0283p3d0cvuglq85log"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "2 1/4 cups all-purpose flour",
                        "1 tsp baking soda",
                        "1 cup butter",
                        "3/4 cup granulated sugar",
                        "3/4 cup brown sugar",
                        "2 large eggs",
                        "2 cups semi-sweet chocolate chips"
                    ]
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Preheat oven to 375°F (190°C)",
                        "Mix dry ingredients",
                        "Cream butter and sugars",
                        "Beat in eggs",
                        "Stir in chocolate chips",
                        "Drop spoonfuls onto baking sheets",
                        "Bake for 9 to 11 minutes"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "Chocolate Chip Cookies"
                },
                "published_at": {
                    "type": "string",
                    "example": "2023-03-10T15:04:05Z"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "dessert",
                        "snack"
                    ]
                }
            }
        },
        "utils.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}